[+] testing complete (golden test)
