+++ .gitignore
@@ -9,3 +9,4 @@
 /out/
 
 *.log
+/src/translator/out/
