+++ src/examples/hello_user_name.ussr
@@ -1,7 +1,12 @@
-текст данные = "Привет, ";
+текст вопрос = "What is your name?";
+текст данные = "Hello, ";
 симв ввод = "x";
+<< вопрос;
+>> ввод;
 пока (ввод != 10){
-    >> ввод;
-    данные + ввод;
+  данные + ввод;
+  >> ввод;
 }
+ввод = "!";
+данные + ввод;
 << данные;
