--- src/examples/math1.ussr
@@ -1,4 +1,0 @@
-цел а = 90;
-цел б = 81;
-цел в = -1* (а * 5 + //б ** 4);  /* -7011 */
-<< в;
