--- src/examples/math2.ussr
@@ -1,12 +1,0 @@
-цел рублей = 5000;
-цел курсБиткойна = 4805444; /* руб за биткойн */
-цел курс2010 = -30 + 45;    /* 15 руб за биткойн */
-цел инфляция = 117 % 1000;  /* процентов */
-цел тысРублейУпущено = 0;   /* тыс рублей могла бы быть сейчас */
-                            /* Это всё просто прикол. Я не серьёзно */
-
-цел рубТогда = рублей * (//10000) / (100 + инфляция);   /* 2304 руб */
-цел биткойнов = рубТогда / курс2010;                    /* 153 биткойна */
-
-тысРублейУпущено = (курсБиткойна / 10 ** 3) * биткойнов; /* 735165 */
-<< тысРублейУпущено;        /* Теперь можно плакать */
