--- src/examples/prob1_simple.ussr
@@ -1,9 +1,0 @@
-цел сумма = 0;
-цел число = 1;
-пока (число < 1000) {
-    если ((число % 3 == 0) | ( число % 5 == 0)){
-        сумма = сумма + число;
-    }
-    число = ++число;
-}
-<< сумма;
