+++ src/examples/prob1s.ussr
@@ -1,0 +1,9 @@
+цел сумма = 0;
+цел число = 1;
+пока (число < 1000) {
+    если ((число % 3 == 0) | ( число % 5 == 0)){
+        сумма = сумма + число;
+    }
+    число = ++число;
+}
+<< сумма;
